+++ internal/restore/workflow_alias.go
@@ -1,6 +1,7 @@
 package restore
 
 import (
+	"github.com/phillipmcmahon/synology-duplicacy-backup/internal/operator"
 	workflow "github.com/phillipmcmahon/synology-duplicacy-backup/internal/workflow"
 	"github.com/phillipmcmahon/synology-duplicacy-backup/internal/workflowcore"
 )
@@ -28,7 +29,7 @@ var DefaultEnv = workflowcore.DefaultEnv
 var DefaultMetadataForEnv = workflowcore.DefaultMetadataForEnv
 var LoadRunState = workflowcore.LoadRunState
 var MetadataForLogDir = workflowcore.MetadataForLogDir
-var NewMessageError = workflow.NewMessageError
-var OperatorMessage = workflow.OperatorMessage
+var NewMessageError = operator.NewMessageError
+var OperatorMessage = operator.Message
 var SignalSet = workflowcore.SignalSet
 var StateFilePath = workflowcore.StateFilePath