+++ internal/workflowcore/request.go
@@ -73,5 +73,3 @@ func (r *Request) Target() string {
 	}
 	return ""
 }
-
-func (r *Request) DeriveModes() {}